+++ CHANGELOG.md
@@ -12,3 +12,9 @@
 - [BREAKING] Bump min Dart SDK version to `3.0.0`.
 - Remove `web` from example.
 - Small refactors & formatting.
+
+## [2.1.0]
+
+- Add `MissingTranslationStrategy` option when creating delegate.
+  (By default it follows the same behavior as originally.
+  You can change it to get the fallback values for missing translations)